Zimbra version match based on major and minor only
Do not check for exact Zimbra version match.

Rationale: Since ZCS 10.0.0 newer versions are being released each two
months as 10.0.1, 10.0.2 and so on. There is not too much change
regarding Java libraries so no need to build again.

In the past we had ZCS 8.8.15 and then the version did not jump even if
we had patches like ZCS 8.8.15p2, ZCS 8.8.15p3 and so on.

In the end ZAL will be built for ZCS 10.0.x, ZCS 10.1.x and so on.